Refactor komand_zendesk -> icon_zendesk. Ignore validator warnings on exception mishandling.
