+++ plugins/zendesk/icon_zendesk/actions/create_ticket/action.py
@@ -97,5 +97,5 @@ class CreateTicket(insightconnect_plugin_runtime.Action):
         try:
             test = self.connection.client.users.me().email
             return {"success": test}
-        except:
+        except:     # pylint: disable=try-except-raise
             raise
