+++ plugins/zendesk/icon_zendesk/actions/delete_membership/action.py
@@ -28,5 +28,5 @@ class DeleteMembership(insightconnect_plugin_runtime.Action):
         try:
             test = self.connection.client.users.me().email
             return {"success": test}
-        except:
+        except:     # pylint: disable=try-except-raise
             raise
