+++ plugins/zendesk/icon_zendesk/actions/delete_ticket/__init__.py
