+++ plugins/zendesk/icon_zendesk/actions/delete_ticket/schema.py
