+++ plugins/zendesk/icon_zendesk/actions/delete_user/schema.py
