+++ plugins/zendesk/icon_zendesk/actions/show_memberships/action.py
@@ -32,5 +32,5 @@ class ShowMemberships(insightconnect_plugin_runtime.Action):
         try:
             test = self.connection.client.users.me().email
             return {"success": test}
-        except:
+        except:     # pylint: disable=try-except-raise
             raise
