+++ plugins/zendesk/icon_zendesk/actions/show_user/action.py
@@ -64,5 +64,5 @@ class ShowUser(insightconnect_plugin_runtime.Action):
         try:
             test = self.connection.client.users.me().email
             return {"success": test}
-        except:
+        except:     # pylint: disable=try-except-raise
             raise
