+++ plugins/zendesk/unit_test/test_suspend_user.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
-from komand_zendesk.actions.suspend_user import SuspendUser
-from komand_zendesk.actions.suspend_user.schema import Input, Output
+from icon_zendesk.actions.suspend_user import SuspendUser
+from icon_zendesk.actions.suspend_user.schema import Input, Output
 from unittest.mock import patch
 
 from insightconnect_plugin_runtime.exceptions import PluginException
